+++ packages/@react-native-windows/cli/src/generator-windows/index.ts
@@ -21,6 +21,7 @@ import {
   copyAndReplaceWithChangedCallback,
 } from '../generator-common';
 import {GenerateOptions} from '..';
+import {CodedError} from '@react-native-windows/telemetry';
 import {findPackage, WritableNpmPackage} from '@rnw-scripts/package-utils';
 
 const windowsDir = 'windows';
@@ -108,15 +109,24 @@ export async function copyProjectTemplateAndReplace(
   options: GenerateOptions,
 ) {
   if (!srcRootPath) {
-    throw new Error('Need a path to copy from');
+    throw new CodedError(
+      'CopyProjectTemplateNoSourcePath',
+      'Need a path to copy from',
+    );
   }
 
   if (!destPath) {
-    throw new Error('Need a path to copy to');
+    throw new CodedError(
+      'CopyProjectTemplateNoDestPath',
+      'Need a path to copy to',
+    );
   }
 
   if (!newProjectName) {
-    throw new Error('Need a project name');
+    throw new CodedError(
+      'CopyProjectTemplateNoProjectName',
+      'Need a project name',
+    );
   }
 
   const projectType = options.projectType;
@@ -245,7 +255,7 @@ export async function copyProjectTemplateAndReplace(
 
   const templateVars: Record<string, any> = {
     useMustache: true,
-    regExpPatternsToRemove: ['//\\sclang-format\\s(on|off)\\s'],
+    regExpPatternsToRemove: [],
 
     name: newProjectName,
     namespace: namespace,